+++ config/initializers/trace_wrapper.rb
@@ -1,0 +1,7 @@
+module TraceWrapper
+  extend NewRelic::Agent::Instrumentation::ControllerInstrumentation
+
+  def self.trace(name)
+    perform_action_with_newrelic_trace(name: name, category: :task) { yield }
+  end
+end