+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.12.2
 - :bulb: Added Zazo::Tools::Logger for logging
+- :bulb: Added newrelic TraceWrapper for cron workers
 - :arrow_up: Updated newrelic configuration 
 
 ### v0.12.1